pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "tailorvision"
version = "0.1.0"
description = "Professional anthropometric measurement extraction from front+side images for garment tailoring."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [{ name = "TailorVision Team" }]
keywords = ["anthropometry", "body-measurement", "smplx", "tailoring", "computer-vision"]

dependencies = [
    # Core numerical / CV
    "numpy>=1.24",
    "opencv-python>=4.8",
    "Pillow>=10.0",

    # Deep learning (body model)
    "torch>=2.0",

    # SMPL-X body model loader
    "smplx[all]>=0.1.28",

    # 2D pose estimation
    "mediapipe>=0.10",

    # Data modelling & serialisation
    "pydantic>=2.0",

    # CLI
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]
bg-removal = [
    "rembg>=2.0",
]

[project.scripts]
tailor-vision = "tailorvision.api.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["tailorvision*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["tailorvision"]
omit = ["tests/*", "tailorvision/__main__.py"]
